LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.clearRect.clip-expected.txt

2d.clearRect.clip

clearRect is affected by clipping regions


PASS clearRect is affected by clipping regions
